move synapse classes out of sub-dirs
