--- bittensor/synapses/dpn/Dockerfile
@@ -1,2 +1,0 @@
-FROM bittensor/bittensor
-COPY examples/cifar bittensor/examples/cifar
