--- bittensor/synapses/ffnn/Dockerfile
@@ -1,2 +1,0 @@
-FROM bittensor/bittensor
-COPY examples/mnist bittensor/examples/mnist
